Добавлена возможность изменять количество товара в счете

+++ domain/src/main/java/com/mwaiterdev/domain/models/response/BillsResponse.kt
@@ -18,7 +18,7 @@ data class BillsResponse(
         @SerializedName("tableGroupId")
         val tableGroupId: Long,
         @SerializedName("tableGroupTotal")
-        val tableGroupTotal: Int,
+        val tableGroupTotal: Float,
         @SerializedName("tables")
         var tables: List<Table>? = arrayListOf(),
         var isExpanded: Boolean = false
